[CELEBORN-297] don't cache file groups for map partition shuffle type
